requirements.txt
#absl-py==2.1.0
#accelerate==1.3.0
#aiohttp==3.11.11
#anyio==4.8.0
#attrs==25.1.0
azure-core==1.32.0
azure-storage-blob==12.24.1
#certifi==2025.1.31
#charset-normalizer==3.4.1
#click==8.1.8
#colorama==0.4.6
#cryptography==44.0.1
#datasets==3.2.0
#deepspeed==0.16.3
docker-pycreds==0.4.0
elasticsearch==8.17.1
fastapi==0.115.8
#filelock==3.17.0
#h5py==3.12.1
#huggingface-hub==0.28.1
#idna==3.10
Jinja2==3.1.5
#keras==3.8.0
librosa==0.10.2.post1
numpy==2.0.2
#onnxruntime==1.20.1
#pandas==2.2.3
#pillow==10.4.0
#protobuf==5.29.3
requests==2.32.3
#scikit-learn==1.6.1
#scipy==1.15.1
#tensorflow==2.18.0
#tensorflow-io-gcs-filesystem==0.31.0
#torch==2.6.0
#transformers==4.48.2
uvicorn==0.34.0
sqlalchemy==2.0.38
pymysql==1.1.1